mysite/requirements.txt
Django==3.2.7
pytest==6.2.5
pytest-django==4.4.0
factory-boy==3.2.0
djangorestframework==3.12.4